refactor: split sessions_patch.rs from sessions.rs to satisfy file-size cap
sessions.rs grew to 641 lines after the chat-session fallback was added
for the event ingest and status-patch paths. Extract the PATCH callback
block (PatchSessionStatusRequest, lifecycle helpers, patch_session_status,
patch_chat_session_status) into a new sessions_patch sub-module declared
via #[path]. sessions.rs drops to 399 lines; sessions_patch.rs is 281.

All test coverage stays: sessions_tests.rs imports from the sibling
sub-module via `use super::sessions_patch::{...}`.
